refactor(ci): name PR-quality modules by responsibility
Drop the enforce-pr-target-* prefix from the split helpers so they match
the existing pr-quality*.cjs naming and describe what they own.
